# user-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  swagger.AttributeResponse:
    properties:
      role:
        $ref: '#/definitions/swagger.RoleResponse'
      status:
        $ref: '#/definitions/swagger.StatusResponse'
    type: object
  swagger.ConflictResponse:
    properties:
      data:
        example: "null"
        type: string
      meta:
        $ref: '#/definitions/swagger.MetaConflict'
    type: object
  swagger.CreateUserRequest:
    properties:
      date_of_birth:
        example: "2000-01-01"
        type: string
      email:
        example: david.afdal@example.com
        type: string
      name:
        example: David Afdal
        type: string
      role_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  swagger.CreatedResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.User'
      meta:
        $ref: '#/definitions/swagger.MetaCreated'
    type: object
  swagger.DeletedResponse:
    properties:
      data:
        example: "null"
        type: string
      meta:
        $ref: '#/definitions/swagger.Meta'
    type: object
  swagger.InternalServerErrorResponse:
    properties:
      data:
        example: "null"
        type: string
      meta:
        $ref: '#/definitions/swagger.MetaInternalServerError'
    type: object
  swagger.Meta:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Success
        type: string
    type: object
  swagger.MetaBadRequest:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Invalid request
        type: string
    type: object
  swagger.MetaConflict:
    properties:
      code:
        example: 409
        type: integer
      message:
        example: user already exists
        type: string
    type: object
  swagger.MetaCreated:
    properties:
      code:
        example: 201
        type: integer
      message:
        example: Successfully created user
        type: string
    type: object
  swagger.MetaInternalServerError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal server error
        type: string
    type: object
  swagger.MetaNotFound:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: user not found
        type: string
    type: object
  swagger.MetaPagination:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total_data:
        example: 2
        type: integer
      total_pages:
        example: 1
        type: integer
    type: object
  swagger.NotFoundResponse:
    properties:
      data:
        example: "null"
        type: string
      meta:
        $ref: '#/definitions/swagger.MetaNotFound'
    type: object
  swagger.ProductsResponse:
    properties:
      created_at:
        example: "2025-03-10 17:31:48"
        type: string
      created_by:
        example: e1107b61-3531-4834-be96-a2cf981bced9
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Product 1
        type: string
      product_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updated_at:
        example: "2025-03-10 17:31:48"
        type: string
      updated_by:
        example: e1107b61-3531-4834-be96-a2cf981bced9
        type: string
    type: object
  swagger.RoleResponse:
    properties:
      created_at:
        example: "2025-03-10 17:31:48"
        type: string
      created_by:
        example: e1107b61-3531-4834-be96-a2cf981bced9
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: admin
        type: string
      updated_at:
        example: "2025-03-10 17:31:48"
        type: string
      updated_by:
        example: e1107b61-3531-4834-be96-a2cf981bced9
        type: string
    type: object
  swagger.SingleUserResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.User'
      meta:
        $ref: '#/definitions/swagger.Meta'
    type: object
  swagger.StatusResponse:
    properties:
      created_at:
        example: "2025-03-10 17:31:48"
        type: string
      created_by:
        example: e1107b61-3531-4834-be96-a2cf981bced9
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: false
        type: boolean
      name:
        example: inactive
        type: string
      updated_at:
        example: "2025-03-10 17:31:48"
        type: string
      updated_by:
        example: e1107b61-3531-4834-be96-a2cf981bced9
        type: string
    type: object
  swagger.UpdateUserRequest:
    properties:
      date_of_birth:
        example: "2000-01-01"
        type: string
      email:
        example: david.afdal@example.com
        type: string
      name:
        example: David Afdal
        type: string
      role_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  swagger.User:
    properties:
      attribute:
        $ref: '#/definitions/swagger.AttributeResponse'
      created_at:
        example: "2025-03-10 17:31:48"
        type: string
      date_of_birth:
        example: "2000-01-01"
        type: string
      email:
        example: david.afdal@example.com
        type: string
      id:
        example: e1107b61-3531-4834-be96-a2cf981bced9
        type: string
      name:
        example: David Afdal
        type: string
      products:
        items:
          $ref: '#/definitions/swagger.ProductsResponse'
        type: array
      updated_at:
        example: "2025-03-10 17:31:48"
        type: string
    type: object
  swagger.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/swagger.User'
        type: array
      meta:
        $ref: '#/definitions/swagger.MetaPagination'
    type: object
  swagger.ValidationErrorData:
    properties:
      date_of_birth:
        example: DoB is required
        type: string
      email:
        example: Email must be a valid email address
        type: string
      name:
        example: Name is required
        type: string
      role_id:
        example: RoleID is required
        type: string
      status_id:
        example: StatusID is required
        type: string
    type: object
  swagger.ValidationErrorResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.ValidationErrorData'
      meta:
        $ref: '#/definitions/swagger.MetaBadRequest'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Documentation of Api for user services.
  title: User Service API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Get paginated list of users
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of users per page
        in: query
        name: limit
        type: integer
      - description: Search query for name or email
        in: query
        name: search
        type: string
      - description: 'Start date format: YYYY-MM-DD'
        in: query
        name: start_date
        type: string
      - description: 'End date format: YYYY-MM-DD'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.UserListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/swagger.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swagger.CreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/swagger.ConflictResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Remove user from database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/swagger.DeletedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SingleUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update existing user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/swagger.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SingleUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
      summary: Update user details
      tags:
      - Users
  /users/{id}/assign-products:
    post:
      consumes:
      - application/json
      description: Assign products to user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
      summary: Assign products to user
      tags:
      - Users
swagger: "2.0"